+++ pdf_bot/bot.py
@@ -10,10 +10,10 @@ from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.ext import Updater, CommandHandler, Filters
 from telegram.ext.dispatcher import run_async
 
-from file import file_cov_handler
-from merge import merge_cov_handler
-from photo_to_pdf import photo_cov_handler
-from watermark import watermark_cov_handler
+from files.file import file_cov_handler
+from commands.merge import merge_cov_handler
+from photos.photo_to_pdf import photo_cov_handler
+from commands.watermark import watermark_cov_handler
 
 load_dotenv()
 APP_URL = os.environ.get("APP_URL")
