+++ pdf_bot/commands/watermark.py
